Move BinaryValue and friends to their own files
